Split LookupService out of api.py
